refactor(BaseMessage): rename 'variant' prop to 'role' and move prop types
